[FIX] product_form_sale_link: assure compatibility with sale_stock
The button 'action_view_sales' is replaced in sale_stock, so it cannot be referenced.

[FIX] product_form_sale_link: replace sales_count by a coherent one

The sales_count field is already used in odoo for another button, which calculates the sum of the quantity of that product for each sale line instead of the number of sale lines.

Now, the number in the button show the correct number of sale lines.

[IMP] product_form_sale_link: black, isort, prettier